reorganize tests into unit/ and integration/ folders, remove redundant files
